fix showing inbox when coming from push / previous load
