+++ shared/actions/config/index.js
@@ -306,8 +306,16 @@ const routeToInitialScreen = state => {
       state.config.startupConversation &&
       state.config.startupConversation !== ChatConstants.noConversationIDKey
     ) {
+      const actions = [
+        RouteTreeGen.createNavigateAppend({
+          path: [
+            {props: {conversationIDKey: state.config.startupConversation}, selected: 'chatConversation'},
+          ],
+        }),
+      ]
       return [
         RouteTreeGen.createSwitchRouteDef({path: [Tabs.chatTab], routeDef: appRouteTree}),
+        RouteTreeGen.createResetStack({actions, index: 1, tab: Tabs.chatTab}),
         ChatGen.createSelectConversation({
           conversationIDKey: state.config.startupConversation,
           reason: state.config.startupWasFromPush ? 'push' : 'savedLastState',
@@ -512,6 +520,7 @@ function* configSaga(): Saga.SagaGenerator<any, any> {
       | RouteTreeGen.ClearModalsPayload
       | RouteTreeGen.NavUpToScreenPayload
       | RouteTreeGen.SwitchTabPayload
+      | RouteTreeGen.ResetStackPayload
     >(
       [
         RouteTreeGen.navigateAppend,
@@ -522,6 +531,7 @@ function* configSaga(): Saga.SagaGenerator<any, any> {
         RouteTreeGen.clearModals,
         RouteTreeGen.navUpToScreen,
         RouteTreeGen.switchTab,
+        RouteTreeGen.resetStack,
       ],
       newNavigation
     )
