+++ shared/actions/route-tree-gen.js
@@ -20,6 +20,7 @@ export const navigateUp = 'route-tree:navigateUp'
 export const putActionIfOnPath = 'route-tree:putActionIfOnPath'
 export const refreshRouteDef = 'route-tree:refreshRouteDef'
 export const resetRoute = 'route-tree:resetRoute'
+export const resetStack = 'route-tree:resetStack'
 export const setInitialRouteDef = 'route-tree:setInitialRouteDef'
 export const setRouteState = 'route-tree:setRouteState'
 export const switchRouteDef = 'route-tree:switchRouteDef'
@@ -35,6 +36,7 @@ type _NavigateUpPayload = void
 type _PutActionIfOnPathPayload = $ReadOnly<{|expectedPath: RCConstants.Path, otherAction: any, parentPath?: ?RCConstants.Path|}>
 type _RefreshRouteDefPayload = $ReadOnly<{|loginRouteTree: RCConstants.RouteDefParams, appRouteTree: RCConstants.RouteDefParams|}>
 type _ResetRoutePayload = $ReadOnly<{|path: RCConstants.Path|}>
+type _ResetStackPayload = $ReadOnly<{|tab: Tabs.AppTab, actions: Array<any>, index: number|}>
 type _SetInitialRouteDefPayload = $ReadOnly<{|routeDef: RCConstants.RouteDefParams|}>
 type _SetRouteStatePayload = $ReadOnly<{|path: RCConstants.Path, partialState: {} | ((oldState: I.Map<string, any>) => I.Map<string, any>)|}>
 type _SwitchRouteDefPayload = $ReadOnly<{|routeDef: RCConstants.RouteDefParams, path?: ?RCConstants.Path|}>
@@ -54,6 +56,10 @@ export const createSwitchTab = (payload: _SwitchTabPayload) => ({payload, type: 
  * ONLY used by the new nav. use this to clear any modal routes
  */
 export const createClearModals = (payload: _ClearModalsPayload) => ({payload, type: clearModals})
+/**
+ * Reset a specific stack. actions is route tree actions TODO better typing
+ */
+export const createResetStack = (payload: _ResetStackPayload) => ({payload, type: resetStack})
 /**
  * Set the tree of route definitions. Dispatched at initialization time.
  */
@@ -77,6 +83,7 @@ export type NavigateUpPayload = {|+payload: _NavigateUpPayload, +type: 'route-tr
 export type PutActionIfOnPathPayload = {|+payload: _PutActionIfOnPathPayload, +type: 'route-tree:putActionIfOnPath'|}
 export type RefreshRouteDefPayload = {|+payload: _RefreshRouteDefPayload, +type: 'route-tree:refreshRouteDef'|}
 export type ResetRoutePayload = {|+payload: _ResetRoutePayload, +type: 'route-tree:resetRoute'|}
+export type ResetStackPayload = {|+payload: _ResetStackPayload, +type: 'route-tree:resetStack'|}
 export type SetInitialRouteDefPayload = {|+payload: _SetInitialRouteDefPayload, +type: 'route-tree:setInitialRouteDef'|}
 export type SetRouteStatePayload = {|+payload: _SetRouteStatePayload, +type: 'route-tree:setRouteState'|}
 export type SwitchRouteDefPayload = {|+payload: _SwitchRouteDefPayload, +type: 'route-tree:switchRouteDef'|}
@@ -94,6 +101,7 @@ export type Actions =
   | PutActionIfOnPathPayload
   | RefreshRouteDefPayload
   | ResetRoutePayload
+  | ResetStackPayload
   | SetInitialRouteDefPayload
   | SetRouteStatePayload
   | SwitchRouteDefPayload
