+++ shared/router-v2/router.shared.js
@@ -4,7 +4,7 @@ import shallowEqual from 'shallowequal'
 import * as RouteTreeGen from '../actions/route-tree-gen'
 import * as Constants from '../constants/router2'
 import * as Tabs from '../constants/tabs'
-import {modalRoutes, routes} from './routes'
+import {modalRoutes, routes, tabRoots} from './routes'
 import logger from '../logger'
 
 const getNumModals = navigation => {
@@ -31,17 +31,6 @@ export const desktopTabs = [
   Tabs.devicesTab,
   Tabs.settingsTab,
 ]
-
-export const tabRoots = {
-  [Tabs.peopleTab]: 'peopleRoot',
-  [Tabs.chatTab]: 'chatRoot',
-  [Tabs.fsTab]: 'fsRoot',
-  [Tabs.teamsTab]: 'teamsRoot',
-  [Tabs.walletsTab]: 'walletsRoot',
-  [Tabs.gitTab]: 'gitRoot',
-  [Tabs.devicesTab]: 'devicesRoot',
-  [Tabs.settingsTab]: 'settingsRoot',
-}
 
 // Helper to convert old route tree actions to new actions. Likely goes away as we make
 // actual routing actions (or make RouteTreeGen append/up the only action)
@@ -108,13 +97,7 @@ export const oldActionToNewActions = (action: any, navigation: any) => {
           ? action.payload.path.last()
           : action.payload.path[action.payload.path.length - 1]
 
-        // a chat, we want inbox/chat
-        if (p === 'chatConversation') {
-          sa.push(NavigationActions.navigate({params: undefined, routeName: 'tabs.chatTab'}))
-          sa.push(StackActions.push({params: undefined, routeName: 'chatConversation'}))
-        } else {
-          sa.push(NavigationActions.navigate({params: undefined, routeName: p}))
-        }
+        sa.push(NavigationActions.navigate({params: undefined, routeName: p}))
       }
 
       // validate sa
@@ -144,6 +127,19 @@ export const oldActionToNewActions = (action: any, navigation: any) => {
         return false
       })
       return isInStack ? popActions : []
+    }
+    case RouteTreeGen.resetStack: {
+      const actions = action.payload.actions.reduce(
+        (arr, a) => [...arr, ...(oldActionToNewActions(a, navigation) || [])],
+        [StackActions.push({routeName: tabRoots[action.payload.tab]})]
+      )
+      return [
+        StackActions.reset({
+          actions,
+          index: action.payload.index,
+          key: action.payload.tab,
+        }),
+      ]
     }
   }
 }
